refactor: extract logic for two-dimensional scatter plot to separate components
